git-svn-id: svn://136.177.114.72/svn_GW/WPhast/trunk@4891 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/MainFrm.h
@@ -64,8 +64,10 @@ protected:
 	DECLARE_MESSAGE_MAP()
 public:
 	afx_msg void OnViewResetDefaultLayout();
-	void UpdateGrid(float x, float y, float z, const char* xy_units, const char* z_units);
-	void UpdateMap(float x, float y, float z, const char* xy_units, const char* z_units);
+// COMMENT: {8/4/2009 5:36:02 PM}	void UpdateGrid(float x, float y, float z, const char* xy_units, const char* z_units);
+// COMMENT: {8/4/2009 5:36:02 PM}	void UpdateMap(float x, float y, float z, const char* xy_units, const char* z_units);
+	void UpdateGrid(double x, double y, double z, const char* xy_units, const char* z_units);
+	void UpdateMap(double x, double y, double z, const char* xy_units, const char* z_units);
 #if !defined(_USE_DEFAULT_MENUS_)
 	afx_msg void OnMeasureItem(int nIDCtl, LPMEASUREITEMSTRUCT lpMeasureItemStruct);
 	afx_msg LRESULT OnMenuChar(UINT nChar, UINT nFlags, CMenu* pMenu);
